federated-intelligence/pyproject.toml
[project]
name = "federated-intelligence"
version = "0.1.0"
description = "MCP server for consulting external AI models (Ollama, OpenAI, Google, OpenRouter)"
requires-python = ">=3.14"
dependencies = [
    "mcp>=1.6.0",
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0"
]

[project.scripts]
federated-intelligence = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
